Login/logout now works.
When item gets added to database it auto-logs out, so we'll need to work on that.

Also need to implement delete_marker.php

+++ ihc_server/admin_server/delete_event.php
@@ -1,14 +1,24 @@
 <?php
-require 'db.php';
-$eventid = $_GET['eventid'];
-$result = $mysqli->query("DELETE FROM ihc_events WHERE eventid='$eventid'");
-if ($result == True) {
-   $message = "Event has been deleted!";
+
+require "./login.php";
+
+if (isset($_SESSION["id"]) && $_SESSION["id"] != null) {
+   require 'db.php';
+   $eventid = $_GET['eventid'];
+   $result = $mysqli->query("DELETE FROM ihc_events WHERE eventid='$eventid'");
+   if ($result == True) {
+      $message = "Event has been deleted!";
+   }
+   else {
+      $message = "Error deleting event!";
+   }
+   $mysqli->close();
+   echo "<script type='text/javascript'>alert('$message');</script>";
+   $url = "../manage_events.php";
+   echo "<script type='text/javascript'>document.location.href = '$url';</script>";
 }
 else {
-   $message = "Error deleting event!";
+   $url = "../admin_auth.php";
+   echo "<script type='text/javascript'>document.location.href = '$url';</script>";
 }
-echo "<script type='text/javascript'>alert('$message');</script>";
-$url = "../manage_events.php";
-echo "<script type='text/javascript'>document.location.href = '$url';</script>";
 ?>
